requirements.txt
fastapi
pydantic
sqlalchemy
psycopg2
psycopg2-binary
uvicorn 
bcrypt
pydantic[email]
python-jose[cryptography]
pyotp 
